Fix `backend-windows` wakers
Move away from `RawWaker`s to reference counted wakers to keep
the tasks window alive as long as wakers for them exist.
In the previous implementation the window could be destroyed
while a waker for it still existed.
In most cases the wakers `PostMessage()` call would return an
error because the target window does not exist anymore.
There is a small chance that after spawning ~65535 more tasks
the window handle is reused and the wrong tasks will be woken.