backend-window: JoinHandle is not `Send`
We unsafely mark the window backend `Task` structure as `Send`/`Sync`
for it to be useable as waker type. While this is correct for the usage
as waker it also made the `JoinHandle` of the window backend `Send`.
Task state access is not thread safe which means that moving and polling
the join handle from another thread is unsound.
Use `PhantomData` with a pointer to make the join handle `!Send`/`!Sync`
again.