Merge backends to get best of both worlds
Remove the hacky window based backend. It was nice as an learning
exercise but going forward the `async-task` crate allows us to improve
the API (less lifetime problems. The previous implementaion probably
contained unsoundness because getting async right is hard.

The new approach uses `async-task` with a window per thread to dispatch
messages. Windows receive messages even when modal loops are running.
We thus do not need to register a hook function to handle this case.

# Conflicts:
#	src/backend/window.rs